requirements.txt
cachetools
numpy
fastapi
pydantic
uvicorn
vosk
nemo_toolkit[asr]